+++ internal/plugin/config.go
@@ -1,8 +1,12 @@
 package plugin
 
 import (
+	"fmt"
+	"io/fs"
+	"strings"
 	"time"
 
+	"github.com/bmatcuk/doublestar/v4"
 	"github.com/meltwater/drone-cache/storage/backend/azure"
 	"github.com/meltwater/drone-cache/storage/backend/filesystem"
 	"github.com/meltwater/drone-cache/storage/backend/gcs"
@@ -37,4 +41,28 @@ type Config struct {
 	SFTP       sftp.Config
 	Azure      azure.Config
 	GCS        gcs.Config
+}
+
+// HandleMount runs prior to Rebuild and Restoring of caches to handle unique
+// paths such as double-star globs.
+func (c *Config) HandleMount(fsys fs.FS) error {
+	mountLen := len(c.Mount)
+	if mountLen > 0 {
+		for i, mount := range c.Mount {
+			if strings.Contains(mount, "**") {
+				// Remove the glob from the original mount list
+				c.Mount[i] = c.Mount[mountLen-1]
+				c.Mount = c.Mount[:mountLen-1]
+
+				globMounts, err := doublestar.Glob(fsys, mount)
+				if err != nil {
+					return fmt.Errorf("glob handle mount error <%s>, %w", mount, err)
+				}
+
+				c.Mount = append(c.Mount, globMounts...)
+			}
+		}
+	}
+
+	return nil
 }